chore(openchallenges): remove ` x-java-class-annotations` from OpenAPI files (SMR-269) (#3372)

+++ apps/openchallenges/mcp-server/src/main/java/org/sagebionetworks/openchallenges/mcp/server/EdamConceptService.java
@@ -67,7 +67,7 @@ public class EdamConceptService {
       description = "Sort field: preferred_label, relevance."
     ) @Nullable EdamConceptSort sort,
     @ToolParam(description = "Sort direction: asc, desc.") @Nullable EdamConceptDirection direction,
-    @ToolParam(description = "List of EDAM concept IDs.") @Nullable List<Long> ids,
+    @ToolParam(description = "List of EDAM concept IDs.") @Nullable List<Integer> ids,
     @ToolParam(
       description = "Free-text search terms to match concept names or descriptions."
     ) @Nullable String searchTerms,
